requirements.txt
absl-py==2.1.0
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==3.2.2
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
asyncio==4.0.0
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
black==24.4.2
bleach==6.1.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.6.15
cffi==1.16.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
configobj==5.0.9
configparser==7.2.0
contourpy==1.2.1
converter==1.0.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
etelemetry==0.3.1
executing==2.0.1
facebook-sdk==3.1.0
fastjsonschema==2.20.0
filelock==3.19.1
fitz==0.0.1.dev2
flake8==7.1.0
flatbuffers==24.3.25
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.7.0
gast==0.5.4
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
google-search-results==2.4.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.34.4
idna==3.7
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
isoduration==20.11.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.4
jiter==0.10.0
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
keras==3.3.3
kiwisolver==1.4.5
langchain-core==0.3.74
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.3
langsmith==0.4.16
libclang==18.1.1
litellm==1.75.9
looseversion==1.3.0
lxml==6.0.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.6.4
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
openai==1.101.0
opt-einsum==3.3.0
optree==0.11.0
orjson==3.11.2
ormsgpack==1.10.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
panda==0.3.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
prov==2.1.1
psutil==6.0.0
pure-eval==0.2.2
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
pyee==11.1.1
Pygments==2.18.0
pylint==3.2.3
pymongo==4.14.1
pyparsing==3.1.2
pypdfium2==4.30.0
pypiwin32==223
pyppeteer==2.0.0
pyppeteer_stealth==2.7.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==2.0.7
pytz==2024.1
pywin32==306
pywinpty==2.0.13
pyxnat==1.6.3
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
rdflib==7.1.4
referencing==0.35.1
regex==2025.7.34
reportlab==4.4.3
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
robotframework==7.3.1
rpds-py==0.18.1
rsa==4.9.1
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
selenium==4.33.0
Send2Trash==1.8.3
simplejson==3.20.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.14.0
tbb==2021.13.0
tenacity==9.1.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.11.0
tinycss2==1.3.0
tokenizers==0.21.4
tomlkit==0.12.5
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
trio==0.30.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20240316
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.4.0
wcwidth==0.2.13
webcolors==24.6.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.3
widgetsnbextension==4.0.11
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.24.0
